+++ src/ui/back/miniquad/mod.rs
@@ -6,11 +6,12 @@
 // - WAIT: [precise input](https://github.com/not-fl3/miniquad/issues/117)
 // - WAIT: [linux resize](https://github.com/not-fl3/miniquad/issues/193)
 
-#[cfg(feature = "alloc")]
-#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "alloc")))]
-mod base;
 mod service; // MiniquadEventHandlerExt, MiniquadService
 mod window; // MiniquadWindow
+
+#[cfg(feature = "alloc")]
+#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "alloc")))]
+mod pixels; // MiniquadPixels
 
 crate::items! { // structural access: _mods, _all
     #[allow(unused)]
@@ -19,7 +20,7 @@ crate::items! { // structural access: _mods, _all
     mod _mods {
         pub use super::{service::*, window::*};
         #[cfg(feature = "alloc")]
-        pub use super::base::*;
+        pub use super::pixels::*;
     }
     pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]